put data exports classes into their own files
