+++ app/services/data_exports/transcription_file_generator.rb
@@ -1,0 +1,183 @@
+require 'csv'
+
+module DataExports
+  class TranscriptionFileGenerator
+    def initialize(transcription)
+      @transcription = transcription
+    end
+
+    def generate_transcription_files
+      files = []
+      files.append(write_raw_data_to_file,
+                      write_consensus_text_to_file,
+                      write_metadata_to_file,
+                      write_line_metadata_to_file)
+    end
+
+    private
+
+    # Private: creates raw data file as tempfile,
+    # returns tempfile
+    def write_raw_data_to_file
+      file = Tempfile.new(["raw_data_#{@transcription.id}-", '.json'])
+      file.write(@transcription.text)
+      file.rewind
+
+      file
+    end
+
+    # Private: creates consensus text file,
+    # returns location of the file
+    def write_consensus_text_to_file
+      file = Tempfile.new(["consensus_text_#{@transcription.id}-", '.txt'])
+      file.write(consensus_text)
+      file.rewind
+
+      file
+    end
+
+    # Private: retrieves and returns consensus text
+    def consensus_text
+      full_consensus_text = ""
+      @transcription.text.each do |key, value|
+        # if we find a frame, iterate through the lines of the frame
+        if /^frame/.match(key)
+          value.each do |line|
+            line_text = line['edited_consensus_text'].present? ?
+                        line['edited_consensus_text'] :
+                        line['consensus_text']
+            full_consensus_text.concat line_text
+          end
+        end
+      end
+
+      full_consensus_text
+    end
+
+    # Private: creates transcription metadata file,
+    # returns location of the file
+    def write_metadata_to_file
+      file = Tempfile.new(["transcription_metadata_#{@transcription.id}-", '.csv'])
+
+      CSV.open(file.path, 'wb') do |csv|
+        transcription_metadata.each do |csv_line|
+          csv << csv_line
+        end
+      end
+
+      file.rewind
+      file
+    end
+
+    # Private: creates transcription line metadata file,
+    # returns location of the file
+    def write_line_metadata_to_file
+      file = Tempfile.new(["transcription_line_metadata_#{@transcription.id}-", '.csv'])
+
+      CSV.open(file.path, 'wb') do |csv|
+        transcription_line_metadata.each do |csv_line|
+          csv << csv_line
+        end
+      end
+
+      file.rewind
+      file
+    end
+
+    # Private: retrieve and return transcription metadata formatted as
+    # array of csv lines
+    def transcription_metadata
+      csv_lines = []
+      csv_lines << [
+        'transcription id',
+        'internal id',
+        'reducer',
+        'caesar parameters',
+        'date approved',
+        'user who approved',
+        'text edited (T/F)',
+        'number of pages'
+      ]
+      csv_lines << [
+        @transcription.id,
+        @transcription.internal_id,
+        @transcription.reducer,
+        @transcription.parameters,
+        @transcription.updated_at,
+        @transcription.updated_by,
+        is_text_edited?,
+        @transcription.total_pages
+      ]
+    end
+
+    def is_text_edited?
+      # iterate through each 'frame' aka 'page' of transcription
+      @transcription.text.each do |key, lines|
+        if /^frame/.match(key)
+          # iterate through each line of the frame
+          lines.each do |line|
+            return true if line['edited_consensus_text'].present?
+          end
+        end
+      end
+
+      false
+    end
+
+    # Private: retrieve and return transcription line metadata formatted as
+    # array of csv lines
+    def transcription_line_metadata
+      csv_lines = []
+      csv_lines << [
+        'consensus text',
+        'line number',
+        'line slope',
+        'consensus score',
+        'line edited (T/F)',
+        'original transcriber username',
+        'line editor username',
+        'flagged for low consensus (T/F)',
+        'page number',
+        'column',
+        'number of transcribers',
+        'line coordinates'
+      ]
+
+      @transcription.text.each_with_index do |(key, value), page_index|
+        # if we find a frame, iterate through the lines of the frame
+        if /^frame/.match(key)
+          page = page_index + 1
+
+          value.each_with_index do |line, line_index|
+            line_number = line_index + 1
+            column = line['gutter_label'] + 1
+            num_transcribers = line['user_ids'].count
+            line_coordinates = {
+              'clusters_x': line['clusters_x'],
+              'clusters_y': line['clusters_y']
+            }
+            line_edited = line['edited_consensus_text'].present?
+            consensus_text = line_edited ? line['edited_consensus_text'] : line['consensus_text']
+
+            csv_lines << [
+              consensus_text,
+              line_number,
+              line['line_slope'],
+              line['consensus_score'],
+              line_edited,
+              line['original_transcriber'],
+              line['line_editor'],
+              line['low_consensus'],
+              page,
+              column,
+              num_transcribers,
+              line_coordinates
+            ]
+
+            csv_lines
+          end
+        end
+      end
+    end
+  end
+end
